fixing dropdown menus
