--- views/device-old.handlebars
@@ -1,12 +1,0 @@
-<div class="text-center">
-  <h2>{{name}}</h2>
-  <p>{{{get_emoji}}}</p>
-  <p>{{description}}</p>
-  <p>Created by {{user.name}} on {{format_date date_created}}</p>
-  <p>
-
-  </p>
-  <p>
-    <button class="btn btn-primary">GadgetGetter</button>
-  </p>
-</div>
