+++ views/device.handlebars
@@ -1,0 +1,17 @@
+{{!-- single device page --}}
+<div class="container">
+    <div class="cardContainer d-flex flex-row flex-wrap">
+        <div class="card" style="width: 12rem">
+                <img
+                    src="images/{{device.name}}.jpeg"
+                    class="card-img-top"
+                    alt="deviceImage"
+                />
+                <div class="card-body">
+                    <h5 class="card-title">{{device.name}}</h5>
+                    <p class="card-text">Type: {{device.device_type}}</p>
+                    <a href="#" class="btn btn-primary">Rent this</a>
+                </div>
+        </div>
+    </div>
+</div>
