+++ views/devices.handlebars
@@ -1,0 +1,6 @@
+{{!-- devices page --}}
+<div class="container">
+    <div class="cardContainer d-flex flex-row flex-wrap">
+        {{> device-details }}
+    </div>
+</div>
