+++ views/layouts/main.handlebars
@@ -160,9 +160,9 @@
                 class="dropdown-menu dropdown-menu-dark text-small shadow"
                 aria-labelledby="dropdownUser1"
               >
-                <li><a class="dropdown-item" href="#">Show All</a></li>
+                <li><a class="dropdown-item" href="/api/devices">Show All</a></li>
                 <li>
-                  <a class="dropdown-item" href="#">Show Available Only</a>
+                  <a class="dropdown-item" href="/api/devices/{{device.is_available}}">Show Available Only</a>
                 </li>
                 <li><a class="dropdown-item" href="#">Show Types</a></li>
               </ul>
@@ -185,10 +185,10 @@
                 class="dropdown-menu dropdown-menu-dark text-small shadow"
                 aria-labelledby="dropdownUser1"
               >
-                <li><a class="dropdown-item" href="#">Show All</a></li>
-                <li><a class="dropdown-item" href="#">IT admins</a></li>
-                <li><a class="dropdown-item" href="#">Teachers</a></li>
-                <li><a class="dropdown-item" href="#">Students</a></li>
+                <li><a class="dropdown-item" href="/api/users">Show All</a></li>
+                <li><a class="dropdown-item" href="/api/users/admins">IT admins</a></li>
+                <li><a class="dropdown-item" href="/api/users/teachers">Teachers</a></li>
+                <li><a class="dropdown-item" href="/api/users/students">Students</a></li>
               </ul>
             </li>
             {{#if logged_in}}
